Create 365_Search_and_Purge.ps1
Requirements:
You may need to enable remote execution, see attached related item "Powershell script will not execute"
You may also need to install the Exchange Online Management module. Before running the script, open powershell and run "Install-Module -Name ExchangeOnlineManagement"

Download the attached .ps1 file to your machine, right click, and run with powershell. 
Click "Enter 365 Admin credentials". This will open a new window to enter 365 admin creds. After entering creds, wait until the status shows as "connected"
Provide a name for the search - can be anything, but cannot be left blank. I would recommend Phish_Search_<date> or something along those lines. Spaces are fine here as well.
Provide a date for when the email was received
Provide email subject. Note -  If the subject of the email you're searching for contains the characters ' or ", they will cause the New-ComplianceSearch command to fail!!!  To work around this issue, double-up the problem character in order to pass it as a string literal to Powershell.  Example - if the subject of the email is This won't be a problem , double up on the apostrophe. It becomes This won''t be a problem. Same goes for quotations, double up each single quotation mark if you have them in the subject.
Provide an email address that the email came from
Should all look like this:
When you're happy with what you entered, hit "start search"
Wait a good 5 minutes or so and then hit "Refresh search status". When the Search status says "Done", you're good to move on to the next step. Keep clicking the "refresh search status" button until the status shows "Done"
When the search is done, click on the "Review Search Results" button. This will display the results of the search in the box to the right. Sorted by the amount of found items for each mailbox. From here, I like to compare this list to what I can see mimecast says was delivered and kind of line them up.
If the list looks good to you, click "I have reviewed the results and am ready to purge" - this unlocks the "Purge" button
Click the "Purge" button to delete the found emails. This also prompts you one last time to make sure you're sure of the results and want to delete the emails. This message pops up in the powershell window, it doesn't pop out at you, so make sure you check the powershell window and enter your response (A to say yes to all)
Purge will start once you confirm in the powershell window
Click "refresh purge status" to check status on the purge. This takes longer depending on the amount of results. If you got a lot of hits, wait 5 or so minutes before refreshing the purge status. 
That's it. This is a very early work in progress. I'm still working to sort out some issues. Feel free to give me (Jake O) feedback on issues or features you would like to see.